+++ slivka/utils/__init__.py
@@ -141,30 +141,6 @@ except ImportError:
             return val
 
 
-def ttl_cache(*, ttl=math.inf):
-    """ Decorator that caches function value for `ttl` seconds.
-
-    The cached value is invalidated after time-to-live.
-    Can only be used with functions having no arguments.
-
-    :param ttl: duration the cached value should remain valid
-    """
-    def decorator(func):
-        invalidate_time = -math.inf
-        cached_result = None
-
-        @functools.wraps(func)
-        def inner():
-            nonlocal invalidate_time, cached_result
-            if invalidate_time < time.monotonic():
-                cached_result = func()
-                invalidate_time = time.monotonic() + ttl
-            return cached_result
-
-        return inner
-    return decorator
-
-
 # noinspection PyPep8Naming
 class class_property:
     """A data descriptor allowing properties on class instances."""
@@ -177,12 +153,20 @@ class class_property:
         return self._getter.__get__(instance, owner)()
 
 
+def alias_property(name):
+    def getter(self): return getattr(self, name)
+    def setter(self, value): setattr(self, name, value)
+    def deleter(self): delattr(self, name)
+    return property(getter, setter, deleter)
+
+
 def deprecated(func):
     @functools.wraps(func)
     def wrapper(*args, **kwargs):
         warnings.simplefilter('always', DeprecationWarning)
+        func_type = type(func).__name__.capitalize()
         warnings.warn(
-            "Function %s is deprecated" % func.__name__,
+            "%s %s.%s is deprecated" % (func_type, func.__module__, func.__name__),
             DeprecationWarning, stacklevel=2
         )
         warnings.simplefilter('default', DeprecationWarning)